internvideo/InternVideo1/Downstream/multi_modalities_downstream/requirement.txt
# python=3.7 required
# might need to use --use-deprecated=legacy-resolver
#-f https://download.pytorch.org/whl/torch_stable.html
bidict==0.21.4
#decord==0.6.0
demoji==1.1.0
docutils==0.18.1
editdistance==0.6.0
einops==0.3.0
ffmpeg_python==0.2.0
imageio==2.9.0
importlib_metadata==4.11.4
numpy==1.19.5
opencv_python==4.6.0.66
pandas
pyarrow
pytorch_lightning
regex==2021.8.28
sacred==0.8.2
scikit_image==0.19.3
scipy==1.7.1
setuptools==58.2.0
#timm==0.4.5
#torch==1.9.0+cu111
torchmetrics==0.9.1
#torchvision==0.10.0+cu111
tqdm==4.64.0
transformers
ftfy
matplotlib